data_indexing/requirements.txt
langchain
langchain-community
pypdf
python-docx
openai
numpy
redis
tqdm